Use FR_tau_2016_vvLoosePresel.root instead of FR_tau_2016.root
The WPs have changed in the new tau ID and the jobs of the sync Ntuple fail because the analyses are unable to access tau FR figures from FR_tau_2016.root. This is a temporary hack anyways, because
a) the tau FRs must be renewed anyways for the new tau ID;
b) until then, we don't compare the event weights b/w sync Ntuples anyways.
Bottom line is that we need the analysis jobs to work.