Fill the correct bin in the cutflow histogram
There was off-by-one error in CutFlowTableHistManager; the label of the last bin was some garbage string b/c it was out-of-boundaries of the input std::vector<std::string> that stores the bin labels. I have no idea how this didn't fail before. Got a runtime error when running 2l2tau analysis in signal region. Was able to reproduce it only when running on a comp node (it ran just fine locally!).