Add first version of the script that inspects sync Ntuples
The purpose of the script is to compare individual objects b/w two sync Ntuples by first dR-matching them in a small cone size (default to 0.01). The comparison is defined in the very last loop of the script, and is problem-defined, i.e. depending on the circumstances, the user has to modify the main loop for their own needs.

TODO: more comments, reasonable variable names, improve printout format, functionality to print only a selection of variables, query for a dR-matched lepton (?), special handling of integer-based variables in the printout.