Reject 2lss1tau flip event if its evt weight 0 due to charge misId
Issue brought up in sync: if an event contains e.g. two muons, or if the combination of charges is such that the charge misidentification probability of a muon must be included to the event weight, the event should be rejected since its event weight is 0 (b/c charge misidentification probability of a muon is 0).