Apply gen-matching correctly for the synchronization
Previously, the MC was too strict: in order for the event to be fully gen-matched,

- the reco leptons (= electrons, muons) were required to be matched to gen prompt leptons (= originating from W/Z/H decays), or from leptonic tau decays
- the reco had taus were required to be matched to gen had taus [*]

within dR < 0.2. However, if we follow the documentation at

https://gitlab.cern.ch/ttH_leptons/doc/blob/fa2a4e95f88bb3118912b9f2e32cfb8369740d0e/2017/MC_matching.md

then reco had taus are also considered to be gen-matched, if they match to prompt gen leptons, or to gen leptons originating from leptonic tau decays. Now that we also consider this additional case, the event is considered to be fully gen-matched if neither reco leptons nor reco had taus aren't matched to gen jets (of course, if gen-matching is enabled and we're working w/ MC and not data). Untested!

There's another problem which I'll fix in the next commit: according to

https://gitlab.cern.ch/ttH_leptons/doc/blob/1b77aa0a44b77f5260bf2fd9a2f4a972e9c6c4ef/2017/event_selection.md

the SRs of 1l+2tau, 2l+2tau, 2lss+1tau and 3l+1tau must include only the events to the sync Ntuple that are fully gen-matched (to 1l+2tau, 2l+2tau, 2l and 3l gen objects, respectively). In the current implementation of the sync workflow, we disregard the gen-matching in the event selection completely.

[*] According to the documentation at

https://twiki.cern.ch/twiki/bin/view/CMS/HiggsToTauTauWorking2015?rev=232#MC_Matching

we should "only consider the decay products of those gen taus that fulfill status flag IsPrompt", which probably means that the gen taus (not its decay products, as they aren't by definition prompt) must be prompt, i.e. originate from W/Z/H decays. However, in GenVisTauProducer,

https://github.com/HEP-KBFI/cmssw/blob/8de4d712912fea9e4c6945a396f7241f3d66845f/PhysicsTools/HepMCCandAlgos/plugins/GenVisTauProducer.cc

we don't apply this kind of check, and unfortunately no MC generator status flags of gen had taus are saved to the nanoAOD Ntuple either. In other words, some of the gen had taus may originate from a hadronic source. I'll probably have to write a module for the GenVisTau collection at nanoAOD-tools level that iterates over the general gen particle collection to find a unique mother and save its PDG ID to a new branch. This has the advantage that we could immediately check the origin of the gen had taus. The check of the promptness comes by requiring the PDG ID of the mother particle to be that of W, Z or H.