Do not require the taus to be gen-matched in 2lss1tau and 3l1tau
According to

https://gitlab.cern.ch/ttH_leptons/doc/blob/1b77aa0a44b77f5260bf2fd9a2f4a972e9c6c4ef/2017/event_selection.md#421-signal-region
https://gitlab.cern.ch/ttH_leptons/doc/blob/1b77aa0a44b77f5260bf2fd9a2f4a972e9c6c4ef/2017/event_selection.md#441-signal-region

the had taus are exempt of gen matching in the SRs of 2lss1tau and 3l1tau channels. I would've expected that apply_hadTauGenMatching is false, when running the 2lss1tau (3l1tau) channels w/ applyFakeRateWeights set to 2lepton (3lepton). Unfortunately, apply_hadTauGenMatching is still set to True in these instances:

https://github.com/HEP-KBFI/tth-htt/blob/f8fccb868505e76abaf09b3470a461eeff3e908b/python/configs/analyzeConfig_2lss_1tau.py#L88
https://github.com/HEP-KBFI/tth-htt/blob/f8fccb868505e76abaf09b3470a461eeff3e908b/python/configs/analyzeConfig_3l_1tau.py#L90

I'm hesitant to flip the booleans here; thus, I explicitly remove the requirement of the had taus to be gen-matched in 2lss1tau and 3l1tau channels.