Create sync plots for all channels
Removed unnecessary data members from NtupleMetaData class: only the file names of the object-level and event-level sync Ntuples are needed; the TTree names are found automatically for each file. The sync plots are created b/w two groups if the TTree names match between two given files. The order matters in which the list of input files are specified. Tested.