Consider reco leptons gen-matched if it matches to a gen had tau
So that we would treat the leptons and taus on equal footing: the taus are considered gen-matched if they match to a gen had tau or a gen prompt lepton or a gen lepton coming from a (gen) tau decay; we now consider leptons gen-matched if they match to a gen prompt lepton or a gen lepton coming from a (gen) had tau decay or a gen had tau. This commit enables the last possibility.