Pick first 2 leptons & taus in 2l2tau channel
This contrasts w/ the previous logic we had, where we simply counted the nof leptons and taus passing certain cuts. E.g. if the event had 2 tight taus and 3 leptons, where the leading and the third lepton passed the tight cuts but the subleading passed only the fakeable cuts, then the event would enter the SR and not fakeable CR. The new logic would categorize this event as fake, b/c not all selected leptons (leading and subleading) are tight. Other consequences are:

- the first two leptons that pass the fakeable leptons selection enter the tight selection -- so that the selected fakeable leptons are a superset of selected tight leptons;
- similar logic applies to taus as well;
- in cleaning the jets and computing the MHT, we use the two selected fakeable leptons (and two fakeable taus);
- the notion of preselected tau is redundant b/c both preselected and fakeable taus are required to pass exactly the same cuts (incl the tau ID WP);
- however, when vetoing events containing more than 2 tight leptons, we are forced to use the full tight lepton collection.