Do not use full tight had tau collection in SR veto of 2l2tau channel
Otherwise, we would reject an event, which contains three (or more) taus, from which either the leading or subleading tau is fakeable, but the rest is tight. Since we always select first 2 leading taus, we would reject this event if we use the full collection b/c the tight tau multiplicity would be 2 (or more), thus vetoing an event that should enter the fake CR and that event wouldn't end up anywhere.