Pick leading lepton & 2 leading taus in 1l2tau channel
See the commit message of c602393e38c262e9ee74833a76d8b8770105085e. Also, removed the two separate selector classes that were used to perform the selections of leading tau and subleading tau separately. If there ever comes a need to apply anti-e and/or anti-mu discriminators on the taus separately, it probably makes sense to add this feature to the selector class (instead of maintining two instances of selectors).