Apply the object selection correctly in 2lss1tau channel
The previous code was wrong: if an event contains two tight leptons and two had taus, from which the leading one is fakeable and the subleading one tight, then this event would enter both SR and fake CR, b/c:

fakeableHadTaus = { leading, subleading }
tightHadTausFull = { subleading }

tightHadTaus = pickFirstNobjects(tightHadTausFull, 1) = { subleading }
fakeableHadTaus = pickFirstNobjects(fakeableHadTaus, 1) = { leading }

=> selHadTaus = tightHadTaus in SR
=> selHadTaus = fakeableHadTaus in fake CR

The remedy is to pick the leading fakeable tau or lepton, and then select only the same object from this collection that passes the tight selection (that's what getIntersection() is for). Finally, the region-specific selected collection is determined from the user-provided cfg parameters.

Tested, but there are some other problems with the object selection. The changes in this channel will soon be implemented in other channels as well.