Count the had taus instead of picking the leading one in 3l/2lss/os1tau
It just occurred to me that the correct way of selecting leptons and taus in a channel depends on the FR method:
- if the object selection is the same (i.e. tight) in both SR and fake CR, we just simply count the number of tight objects and check if it satisfies the required multiplicity set by the channel;
- if the object selection varies b/w SR and fake CR (i.e. tight in SR, fakeable in fake CR), we cannot simply count the nof objects and check their multiplicity, instead we have to always pick the first N leading objects and check if they satisfy the multiplicity cuts set by channel.

Example #1: in 2lss1tau channel, the FR method is applied only to the leptons. This in turn means that the taus must be tight in both SR and fake CR. Therefore, we can simply count the nof tight taus to see, if the event at hand is a valid 2lss1tau event. Events which potentially overlap w/ 3l1tau SR are vetoed by requiring the nof tight leptons be no more than 2; the events which overlap w/ 2l2tau are vetoed by requiring that no more than 1 tau passes Medium tau ID WP (this is one notch higher WP than what is used in other channel to define a tight tau). Similar logic holds for 2los1tau and 3l1tau channels.

Example #2: in 1l2tau channel, the FR method is applied to both leptons and taus. This means that in fake CR the taus may only pass the fakeable definition. However, we cannot simply count the nof fakeable or tight taus to see if they would enter the fake CR or SR. For instance, an event contains 1 tight lepton and 4 taus, from which the first 2 leading pass the fakeable cuts and the last 2 leading pass the tight cuts. If we were to count the nof tight taus, we would classify this event as 1l2tau SR event; if we were to pick only the first two leading 2 taus but still count the nof tight taus in the event, this event would be rejected completely b/c of SR veto. Thus the only way we can keep things consistent is to always pick the first 2 leading taus and check and impose SR veto in fake CR by checking if they both pass the tight cuts. Similar argument holds for 2l2tau. Since other channels veto 1l2tau and 2l2tau events, there will be no overlap b/w SRs.

Tested. Got closure in 2lss1tau and 3l1tau channels.