Implement RLE Venn diagram from the events chosen in all channels
This is needed for self-validation purposes, to make sure that

a) we don't select the same event in two or more SRs;
b) that an event we select in SR is not selected in complementary fake (or flip) CR, and vice versa.

Also, implemented the same logic for every unique pair of sync Ntuples, to see how the events migrate b/w two groups. E.g. it's now easier to see that some events we select in 1l2tau SR are acutally selected in 1l2tau fake CR by another group.