Add 'missing_from_superset' key to lepton FR dictionaries
This key is needed only in Ntuple production and can be omitted in dictionaries that are used in the analysis (i.e. dictionaries w/o nanoAOD in the file name). Plus, this information is required only in data samples, which means that there's no need to include it to the sync Ntuple dictionary.