Collect common py cfg options to the main base class
Now, all analysis channels (except for some auxiliary analyses) automatically write analysis options to python cfg files if
a) the option is specified in the local variable jobOptions_local in analyzeConfig.py; and
b) the option is defined in jobOptions dictionary in individual derived classes.
Uncommon options must be passed via additionalJobOptions parameter (as is currently done in inclusive, jetToTauFakeRate and LeptonFakeRate analyses).

Also, provided the CLI option to enable HLT trigger object matching in ttW, ttZ and WZ CRs and in jetToTauFakeRate analysis. Also, added this option to the sync as well. HLT trigger object matching is currently disabled everywhere, but we have the functionality to turn it on at any point if we want to.