Update the files containing sums of LHE weights for THQ & THW
The files are generated by running comp_lhe_weightsum.py on the MINIAODs produced by Marco:

comp_lhe_weightsum.py \
  -i /hdfs/local/karl/store/mc/private/THW_5f_Hincl_13TeV_madgraph_pythia8_Fall17 \
     /hdfs/local/karl/store/mc/private/THQ_4f_Hincl_13TeV_madgraph_pythia8_Fall17 \
  -o data/LHEWeightSum/THW_5f_Hincl_13TeV_madgraph_pythia8_Fall17.txt \
     data/LHEWeightSum/THQ_4f_Hincl_13TeV_madgraph_pythia8_Fall17.txt

The first column denotes the weight ID and the second column is the sum of all weights. According to dump of LHE header and these links:

- https://github.com/cms-sw/genproductions/blob/master/bin/MadGraph5_aMCatNLO/cards/production/2017/13TeV/Higgs/thw_5f_ckm_LO_ctcvcp_MH/thw_5f_ckm_LO_ctcvcp_MH_reweight_card.dat#L25
- https://github.com/cms-sw/genproductions/blob/master/bin/MadGraph5_aMCatNLO/cards/production/2017/13TeV/Higgs/thq_4f_ckm_LO_ctcvcp_MH/thq_4f_ckm_LO_ctcvcp_MH_reweight_card.dat#L25

the LHE weight that corresponds to SM is rwgt_12, the sum of which will be used to recompute process cross section (by dividing the total cross section by the sum of SM weights; since both privately produced samples are inclusive, no BRs won't be taken into account as was the case in previous analysis).