Disable automatic resubmission by default
The problem is that when everyone start producing the datacards and the jobs start to fail (due to simple I/O errors on /hdfs), the jobs will die. This is caught by sbatchManager, which then raises an exception and exits. Then the makefile notices that one target exited w/ a non-zero return code and exits as well, although the SLURM jobs are still running. After that the automatic resubmission kicks in, which creates new jobs if the target files do not exist. However, the targets may very well be produced by running SLURM jobs, which results in duplicate jobs. This can cascade even more if the first resubmission fails.

For a short term solution I think it's preferable that the automatic resubmission is disabled by default. If the user really wants to resubmit automatically N times, they can do so by specifying -r N option for tthAnalyzeRun scripts. Eventually, this resubmission business must be integrated into sbatchManager, b/c that's where we bookkeep jobs that are still running.