Remove jet cleaning completely in produceNtuple
Since the jet cleaning has now become channel-dependent (where the required number of fakeable objects varies channel-by-channel), it makes sense to remove the jet cleaning completely in produceNtuple (i.e. post-processing) step. Previously, we cleaned w.r.t the whole fakeable lepton collection, which might explain some of the differences we in the sync w/ the other groups. The only downside this change might have is that we increase the final Ntuple size a bit.