Save the status of the software to a text file whenever running jobs
The text file is generated once every time someone (re)submits the jobs. The text file contains the following information:
- last 3 commits in a repository;
- status of the repository (i.e. are there modified or untracked files)
- diff on the tracked files against HEAD
The selection of repositories is different for each type of workflow (the motivation begin, we don't need to know in which state the MEM repositories are when generating the PU profiles). The log file is saved to the cfg directory (on /home) and to the output directory (usually /hdfs) if it's different from the cfg directory.

Tested jobTools.record_software_state() function individually, but haven't tested the whole thing, yet.