Recategorize VHToNonbb and ST_tWll samples to VH and TT
It's not clear, yet, to which category VHToNonbb should belong to -- additional_signal_overlap or a separate VH category. For now we opt for the latter, b/c there might be a sizeable contribution from this process to the signal region. (Need to follow up w/ Marco on this one.)