Add labeled histograms bin-by-bin correctly
Added a function which checks if all bins in [1, GetNbinsX()] are labeled or not; if they are, then now underflow or overflow bins are added. If they aren't then the default behavior applies (i.e. both underflow and overflow bins are summed).

This fixes a problem where the number of bins in MEtFilterCounter MC histograms doubled compared to the original count. As this wasn't the case for data (which I don't fully understand), the procedure failed where prompt MC is subtracted from data.

Tested.