Make checking the output files optional (as opposed to input files)
Since the sbatchManager internally checks the validity of the output file during the jobs submission anyways, it makes sense to provide an option for the user to disable the same validation procedure when resubmitting the whole workflow from scratch again (by rerunning the tthAnalyzeRun scripts). This can be done via -C flag -- the config files are regenerated for the targets in the makefile that do not exist on disk. This speeds up config file generation by orders of magnitude. Tested.