Clean presel electrons wrt presel muons in produceNtuple
This bug didn't affect the preselected electrons b/c in the analysis code we already do the right thing (clean presel electrons wrt presel muons). The only consequence it has is that the taus are cleaned wrt the preselected electrons and preselected muons, meaning that previously we may have cut out a bit more taus than necessary, but this is a very small effect (we can lose a tau only if it matches to a preselected electron within dR=0.3 that is not cleaned wrt preselected muons (that fail fakeable selection) within dR=0.05).