Increase the basket size of produceNtuple jobs from 16kb to 32kb
This change helps to combat against the problem so far occurring only in produceNtuple jobs: the memory consumption may occasionally hit the memory limit of 1800M set by the SLURM jobs. Doubling the basket size of all branches decreases the peak memory consumption to 300M, but also increases the runtime by 30%. Since we never experienced any memory-related problems in the analysis jobs, we'll keep the basket size to default there.