Propagate the number of objs from reco writers to gen writers
It turns out that the issue described in https://github.com/HEP-KBFI/tth-htt/issues/36 originated from the fact that the number of reco jets exceeded the number of matched gen jets. The number of gen objects in the reco writers is now propagated to the associated gen writers. Also, made sure that the object multiplicities are equal b/w the same type of reader and writer classes.