Split ttbar to estimate irreducible bkg and for BDT training in 0l2tau
Unlike in other channels, the two taus in ttbar entering the SR of 0l2tau channel are likely real taus, which means that there's a significant irreducible contribution of ttbar events to the SR of 0l2tau channel. This has the effect that we cannot use both PSweights samples (which is higher in statistics) and non-PSweights samples (lower in statistics) in both the analysis and in BDT training, which in turn means that we have to reserve non-PSweights samples to estimate irreducible contribution to the SR and PSweights samples for BDT training. In other channels we can use both sets of samples in BDT training and in the analysis to estimate shape uncertainties of tau fake rates.