Delay the summation of eff event counts until py cfg files are generated
Previously, the effective number of event counts was done during the creation of sample dictionaries. However, in 0l2tau channel (and probably in some other future cases as well) we want to treat TT PSweights and TT non-PSweights samples on separate footing. This means that we sometimes need to sum their effective event counts (in all other channels) but sometimes we have to keep the event counts separate (like in 0l2tau channel).

From now on all sample dictionaries must contain a key 'sum_events' which lists the samples that cover the same phase space and thus their effective event counts must be summed in order to produce the correct lumiscale weight. Whene new analysis jobs are created, the effective event counts are summed (in analyzeConfig) before the loop over sample entries kicks in (in create() function). This means that 'sum_events' key must be skipped when looping over the entries in sample dictionary.