Possibility to require taus to pass OR of two tau ID WPs
In HH analysis we're not sure, yet, which tau ID WP we're going to use. In order to test both WPs, the tau collection in the post-processed Ntuples must pass the loosest WP of dR03mva or dR05mva. Previously, the FW supported only a single tau ID WP; but now it's possible to require the taus to pass either a dR03mva WP or a dR05mva WP. Both WPs must be joined by an ampersand, e.g.:

dR03mvaLoose&dR05mvaVLoose

means that the taus are required to pass either dR03mvaLoose WP or dR05mvaVLoose WP. This type of cut can be applied to any collection (whether it be loose, fakeable or tight collection).

However, if the cut is applied to tight collection, but the WPs are looser than what is nominally assigned to loose & fakeable collection, we would still lose those taus since the loose/fakeable cut on tau ID is tighter than that for the tight collection. For this reason I added a relatively automatic way of resetting those WPs to the new ones in the loose and fakeable tau selectors if the new WPs are somehow looser than what was initially assigned to them. Rules are explained in src/RecoHadTauSelectorBase.cc, lines 249-256.