Add secondary protection when reading the subjets of AK8 fatjets
There seems to be an issue in CMSSW/nanoAOD framework, b/c the event
305112:442:734209267 in Tau_Run2017F_17Nov2017_v1 dataset has reportedly
AK8 subjets according to FatJet_subJetIdx1 and FatJet_subJetIdx2
branches, but there are actually no Subjet objects in the same event.
This inconsistency needs to be investigated and probably followed up with
the CMSSW/nanoAOD devs. For now, we also check how many subjets have
been read before we link subjets with a FatJet.