Correctly fill the 1st event in the output Ntuple when post-processing
The error that this commit attempts to fix was introduced here: https://github.com/HEP-KBFI/tth-htt/commit/1aef5a767ee2807b97132b5b14526766de3f146d#diff-4b03f20f4bc9b4e342e838e027f53cd3
When we migrated from TChain to TTreeWrapper in produceNtuple, the following happened:
- the TTreeWrapper was initialized, which basically meant
    - gathering the list of input files
    - recording the *Reader classes
    - recording the basket size of the input TTree
- hasNextEvents() was called, which
    - opened a new file if there was none open at the time
    - initialize the TTree if there was none present at the time
    - set the basket size
    - set the branch addresses in the *Reader classes
    - read the event via GetEntry()
- the association was created for the branches between the input and output TTree by setting the branch addresses to the local variables
- the output TTree was filled with the new objects returned by the *Reader classes as well as the local variables that were bound to the input TTree

However, in the second-to-last step the local variables remained to their default values since SetBranchAddress() was called *after* GetEntry() on the input TTree. In other words, all variables that were supposed to be copied over automatically were actually set to 0 in the very first event in every produceNtuple job. This bug affects ~80k events in total.

The solution here is a bit hack-y, but it works: the hasNextEvent() call doesn't call GetEntry() on the very first event until it's explicitly called in every produceNtuple job. This way we can still get a valid TTree that we can bind to the local variables.

Tested, but not very thoroughly.