Do not require samples subject to stitching to be present in sample dict
We had defined stitching relations for the two parts of W+jets samples
but these samples are not present in any sample dictionary other than
the sample dictionary used in HH 2lss analysis. Decided to replace the
assertion statements as they were previously used to guide the
implementation of sample stitching.