Use common logging module w/ preset format across the whole project
The problem is that if a module (in our current case it's the hdfs module) makes the first logging statement (e.g. logging.debug()) and the config of the logging module is not set up, then any subsequent attempt to set up its config will fail. This peculiarity can be fixed only by setting the config after the first import of the logging module. This, in turn, implies that all python files import logging module with the same config. That's why we now have one "central" logging module that all files import from.