Hardcode namenode and port & use complete URI in hdfs module
There *may* be some overhead in resolving the namenode name and port number. By hardcoding these two to their actual values (as opposed to letting the libhdfs perform DNS-like action of finding the proper namenode name and port number) as well as using proper URI in queries to the libhdfs library, we may gain a bit in performance.