Update comments on HLT paths & update average prescales for 2017
The prescale factor for HLT_Mu20 was off by 3x: it was 225 but should be 72. Since the prescale factor for HLT_Mu27 is also 225, I suspect it to be copy-paste error.
It also looks like the prescale factors for the following HLT paths were computed with the assumption that all these paths were enabled in all acquisition eras but were in fact disabled in acquisiton era B:
- HLT_Mu3_PFJet40
- HLT_Ele8_CaloIdM_TrackIdM_PFJet30
- HLT_Ele17_CaloIdM_TrackIdM_PFJet30
- HLT_Ele23_CaloIdM_TrackIdM_PFJet30
I updated numbers w/ the same assumption but I'm now wondering if that's the right thing to do -- if the HLT path is excluded for a particular acquisition era, shouldn't it be taken into account in the computation of the prescale factor?