Take the whole run into account when computing prescale factors
Previously considered only the acquisition eras when the HLT path was enabled when computing total integrated luminosity. However, it makes more sense to interpret the HLT path as highly prescaled if it's not been enabled for some time of the run even though it ran unprescaled when it was enabled. The numbers in LeptonFakeRate_trigger_cfi do not need to be changed as they already have been computed w/ the new logic in mind.