Do not remove output file in the makefile of sync workflow
The check is done already in SLURM jobs: if the file exists, then the job is skipped; if it doesn't then the job is run. However, previously the job was always run b/c the output file never existed b/c it was removed in the makefile just before the job was submitted.