Do not remove hadd targets before the job is being run
The problem is that when the makefile is run multiple times, all hadd results from the previous run are deleted. This is unnecessary b/c the check of file existance is done in the hadd jobs, not at the makefile level. Therefore, no files should be deleted b/c rerunning make would delete all previous results of hadd jobs.