Do not use hdfs module in get_log_version
Using the hdfs module there has very little impact on the performance -- only the version log files that would require.. version numbers are stored on /hdfs; the rest of the log files reside on /home so it makes sense to use POSIX commands directly to check if the files are present on the file system.

Moreover, when an instance of ClusterHistogramAggregator is created, it creates an instance of sbatchManager, which in turn import get_log_version (because we want consistent versioning of the log files). Current design of the makefile creates multiple instances of ClusterHistogramAggregator and thereby multiple instances of JVM via the hdfs module. This causes problems because the system runs out of resources to create the JVMs. The only solution is to completely remove any calls to the hdfs module in sbatchManager.

Tested, didn't see any lines in the verbose output of ClusterHistogramAggregator indicating the instantiation of the hdfs interface.