Read only the chosen sys unc in the analysis jobs
In Ntuple production we want to read all branches that are affected by certain systematic uncertainties, but in the analysis job we want to read only those branches that are specified in the analysis config file. This is because reading all branches is time-consuming and if we only need a fraction of infromation that is available in the Ntuples, we would want to keep the number of branches to a minumum that we bind to.
Untested (not sure if it even compiles).