Disable electron cleaning by default in produceNtuple
The taus are still cleaned wrt the electron collection that's been cleaned wrt the muons -- to maintain consistency between produceNtuple and the analysis.
For more context, read https://github.com/HEP-KBFI/tth-htt/issues/67
Compiles & tested on 2017 sync Ntuple. The number of preselected electrons jumped from 23473 to 23607, but the number of preselected taus remained the same (49591).