Drop FR systematics in the tight regions
The only exceptions are channels with taus when the FR method is applied only to the leptons; in this case the difference b/w data & MC is compensated with SFs. Therefore, associated systematic uncertainties need to be preserved, even in the SR when running the analysis on MC events. Currently, this is the case for 2lss+1tau and 3l+1tau channels, but the support for enabling/disabling data/MC SFs and associated systematics is developed for other channels, too.

The code also disables implicitly the systematics that are not used in the given FR model. For instance, in pure leptonic channels there's no need to apply jet->tau FRs or associated SFs because no taus are selected in this channel. Therefore, systematics relevant to jet->tau FR are also disabled in pure leptonic channels.

Tested all channels and CRs. See https://github.com/HEP-KBFI/tth-htt/issues/64 for more.