Use nominal tau ID WP in applying data/MC SF when relaxing the WP
When producing Ntuples for the BDT training, we want to relax the tau ID WP of the tight tau collection to gain more statistics. The difference in the event yields in the relaxed conditions is compensated with data-to-MC SF to match the event yields in the SR. However, this means that the nominal tau ID WP in the SR must match to the WP used in the numerator of the SF, which was not always the case, as sometimes the nominal tau ID WP and the WP of the SF numerator were different. The solution is to never override the nominal tau ID WP in if-block that relaxes the tau ID WP.
https://github.com/HEP-KBFI/tth-htt/issues/68