Ntuples with event-level preselection
The Ntuples are preselected by requiring each event to contain at least 2 fakeable electrons or muon or hadronic taus that pass the loosest WPs of either 2017v2 MVA or DeepTau ID discriminator. There are absolutely no requirements on the jets. The events are also required to have at least one HLT path triggered that we use in the analysis.
https://github.com/HEP-KBFI/tth-htt/issues/61