Create EvtWeightRecorder class & record some weights in 2lss+1tau
WORK IN PROGRESS.
Managed to encapsulate all event level weights that depend on the choice of systematics except for the ones that vary jet->lepton FR, jet->tau FR and DY reweighting weights.
The idea is to provide a separate list of systematics that is processed internally ('central_or_shift_local'). The list of systematics is passed to the job iff the job runs on the central values.
Also, the filling of histograms needs to be figured out..

First commit of many to solve this task: https://github.com/HEP-KBFI/tth-htt/issues/60
Compiles, otherwise completely untested.