Run dedicated systematics jobs for ctcvcp samples
Since the central jobs that run most of the systematics (not affecting the event selection) internally still consumes too much memory (4.2GB) after optimizing for the histograms, the only solution is to run dedicated jobs for all systematics in ctcvcp samples. We'll keep the optimization for histograms, because otherwise the hadd jobs would fail.