Remove memory cap of 1800M from the SLURM jobs
It has no effect anymore: the nodes are configured to handle jobs that consume up to 2048M of memory per core, and the jobs are accepted depending on the CPU load or memory consumption in the comp nodes dynamically. The limit 1800M originated from the days where the number of cores wasn't forced down to a lower number in comp nodes that had less than 2GB memory per core, so the solution was to lower the memory cap for all jobs, regardless of which comp node they were run.