Log contents of working directory
A random error appeared during the hadd stage of 2lss+1tau analysis: hadd claimed that the output file already exists, but it should not be possible by construction, as the current working directory is unique to each SLURM job. Maybe increasing the logging will shed some light on the matter.