Fix error handling when reading back MEM scores
It's not always possible to compute the MEM score in the selected events for valid reasons. For instance, if the event is tH-like, it doesn't satisfy the event level cuts applied in MEM computation. Therefore, these errors must be propagated to the output Ntuple and handled in the analysis code. The only sane way is to save a dummy MEM object with non-zero error code. The error handling in the analysis code was flawed, though, because we tried to read the error code of matched MEM object. Because there are no such matches, error code of 0 was read back because the MEM object remained unitialized.