Do not split W+jets into two parts in 3l, 2lss+1tau (default, BDT)
The problem is that the process category name "WZ" not only includes WZTo3LNu dataset (which we want to split into half: one half for the BDT training, the other half for the regular analysis), but also W+jets samples as well. The best way to go around this issue is by attaching a number to every sample which we want to split between training and analysis. This number ('skipEvery'), call it N, is specified in tthAnalyzeRun and has the following functionality:
- when running in BDT mode, every N-th event is *skipped* and every other (N-1) event is retained, whereas the event weight is scaled by N / (N - 1);
- when running in analysis mode, every N-th event is *kept* and scaled by N, whereas every other (N-1) event is skipped.
Tested locally, seems to work.
Related: https://github.com/HEP-KBFI/tth-htt/issues/91