Show systematic uncertainties in the prefit plots
The statistical uncertainties are rebinned *after* the bin-by-bin uncertainties are found from un-rebinned histograms. The correct approach would be rebinning all histograms (including the ones for each systematic shift) and then derive the systematic uncertainties. However, this is a substantial amount of work and I'm unfamiliar with the code... The current approach is a bit better than not showing the systematic uncertainties at all.
Tested.